fix(cli,cognify): collapse colliding point ids, and dispatch vector-reindex through the trait
Review findings on #231.

**edge-reindex wrote the same point id twice in one batch.** `counts_by_text`
keyed on the raw retrieval text, but the id runs that text through
`normalize_identifier` (lowercase, spaces to underscores, apostrophes dropped),
and nothing upstream normalises `relationship_name` before it reaches the graph
— `expansion.rs` stores the raw LLM string. So "works at", "Works At" and
"works_at" were three map entries and one id. The per-run writer never had to
care because it only ever holds one run's edges; this is the first path that
aggregates over the whole graph, which is what makes the collision reachable.

Colliding texts sort adjacently, so they usually land in the same batch, and
pgvector writes a batch as one multi-row `INSERT ... ON CONFLICT (id) DO
UPDATE`. Postgres rejects a repeated row outright (SQLSTATE 21000), so
`--apply` aborted the whole run with an opaque "Vector DB error" having written
nothing for that batch. LanceDB did not error but silently kept one row, so
`points_written` over-reported and `number_of_edges` carried a single
spelling's share. Collapse by id before building the batch, keeping the first
text in cursor order and summing the counts.

**vector-reindex gated on the provider string.** The trait method was added
with an `Ok(0)` default precisely so the CLI need not know its backend, and the
command then compared `vector_db_provider != "pgvector"` and claimed the
backend had "no ANN index to backfill". That is asserted, not checked: an
out-of-tree adapter registered through the component registry can override the
method, and was being told it has no repair path — the gap the command exists
to close. Dispatch unconditionally and branch on the returned count, naming the
provider only in the zero case.

**`--limit 0` was a dead end.** A zero budget breaks out of the write loop
before `resume_cursor` is set, so the run warned "re-run with the reported
cursor" while reporting none. Rejected at parse time instead.

Also corrects a doc comment claiming the cognify writer calls `point_id_for`;
it still reaches the same value through `edge_retrieval_text` +
`new_deterministic`.

The collision test fails without the fix (4 distinct texts, not 2). The
vector-reindex e2e case was pinning the provider guard, so it now pins the
post-dispatch summary instead.

| Check | Exit |
|---|---|
| `cargo fmt --check` | 0 |
| `cargo clippy --all-targets -- -D warnings` | 0 |
| `cargo test -p cognee-cognify --test edge_reindex_backfill` (11) | 0 |
| `cargo test -p cognee-cli --test cli_e2e` (36) | 0 |

+++ crates/cli/src/cli.rs
@@ -318,8 +318,10 @@ pub struct EdgeReindexArgs {
     /// Bounds writes only, so it has no effect without `--apply` — a report
     /// writes nothing to cap. It does not bound the graph read, which is
     /// whole-graph either way.
-    #[arg(long = "limit")]
-    pub limit: Option<usize>,
+    /// A limit of `0` is rejected rather than silently doing nothing: it would
+    /// report no cursor to resume from, so there would be no way to continue.
+    #[arg(long = "limit", value_parser = clap::value_parser!(u64).range(1..))]
+    pub limit: Option<u64>,
 
     /// Resume an interrupted run: skip every retrieval text up to and including
     /// this one. Pass the cursor the previous run reported.
